refactor: 오디오 관련 패키지 분리
