+++ src/main/java/com/jolupbisang/demo/application/audio/dto/StepFunctionOutput.java
@@ -1,0 +1,20 @@
+package com.jolupbisang.demo.application.audio.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
+
+public record StepFunctionOutput(
+        String statusCode,
+        List<MergedPath> paths
+) {
+
+    public record MergedPath(
+            @JsonProperty("user_id")
+            Long userId,
+
+            @JsonProperty("s3_path")
+            String s3Path
+    ) {
+    }
+}
