+++ src/main/java/com/jolupbisang/demo/application/segment/SegmentService.java
@@ -8,13 +8,13 @@ import com.jolupbisang.demo.application.segment.dto.SocketSegmentRes;
 import com.jolupbisang.demo.domain.meeting.Meeting;
 import com.jolupbisang.demo.domain.segment.Segment;
 import com.jolupbisang.demo.domain.user.User;
-import com.jolupbisang.demo.global.response.SocketResponse;
-import com.jolupbisang.demo.global.response.SocketResponseType;
 import com.jolupbisang.demo.infrastructure.meeting.MeetingRepository;
 import com.jolupbisang.demo.infrastructure.meeting.audio.AudioProgressRepository;
 import com.jolupbisang.demo.infrastructure.meeting.client.dto.response.DiarizedResponse;
 import com.jolupbisang.demo.infrastructure.segment.SegmentRepository;
 import com.jolupbisang.demo.infrastructure.user.UserRepository;
+import com.jolupbisang.demo.presentation.audio.dto.response.SocketResponse;
+import com.jolupbisang.demo.presentation.audio.dto.response.SocketResponseType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
