--- src/main/java/com/jolupbisang/demo/global/response/SocketResponse.java
@@ -1,10 +1,0 @@
-package com.jolupbisang.demo.global.response;
-
-public record SocketResponse<T>(
-        SocketResponseType type,
-        T data
-) {
-    public static <T> SocketResponse<T> of(SocketResponseType type, T data) {
-        return new SocketResponse<>(type, data);
-    }
-}
