--- src/main/java/com/jolupbisang/demo/global/response/SocketResponseType.java
@@ -1,5 +1,0 @@
-package com.jolupbisang.demo.global.response;
-
-public enum SocketResponseType {
-    LAST_PROCESSED_CHUNK_ID, ERROR_MESSAGE, DIARIZED_SEGMENT
-}
