+++ src/main/java/com/jolupbisang/demo/presentation/audio/dto/request/SocketRequestType.java
@@ -1,0 +1,5 @@
+package com.jolupbisang.demo.presentation.audio.dto.request;
+
+public enum SocketRequestType {
+    CANCEL_COMPLETION,
+}
