--- src/main/java/com/jolupbisang/demo/presentation/meeting/dto/request/SocketRequestType.java
@@ -1,5 +1,0 @@
-package com.jolupbisang.demo.presentation.meeting.dto.request;
-
-public enum SocketRequestType {
-    CANCEL_COMPLETION,
-}
